requirements.txt
streamlit==1.40.0

torch==2.9.0
torchvision==0.24.0

numpy==2.3.2
pillow==11.0.0

joblib==1.3.2
protobuf==4.25.1
typing-extensions>=4.8.0
